refactor(acceptance): extract helpers to acceptance-helpers.ts
Move isStubTestFile, isTestLevelFailure, loadSpecContent,
loadAcceptanceTestContent, buildResult, regenerateAcceptanceTest,
and _regenerateDeps from acceptance-loop.ts to a new
acceptance-helpers.ts file.

Also remove the unreachable legacy fallback path
(generateAndAddFixStories + executeFixStory) — the strategy enum
is "diagnose-first" | "implement-only" only, so the fallback was
dead code.

acceptance-loop.ts: 1051 → 720 lines
acceptance-helpers.ts: 222 lines (new)
