+++ src/execution/lifecycle/acceptance-loop.ts
@@ -9,37 +9,44 @@
  * 5. Retries until max retries or all tests pass
  */
 
-import path, { join } from "node:path";
-import { type FixStory, convertFixStoryToUserStory, generateFixStories } from "../../acceptance";
 import { loadAcceptanceTestContent as loadAcceptanceTestContentModule } from "../../acceptance/content-loader";
 import { diagnoseAcceptanceFailure } from "../../acceptance/fix-diagnosis";
 import { executeSourceFix } from "../../acceptance/fix-executor";
 import { loadSemanticVerdicts } from "../../acceptance/semantic-verdict";
 import {
   findExistingAcceptanceTestPath as findExistingAcceptanceTestPathFromOptions,
-  resolveAcceptanceFeatureTestPath,
   resolveAcceptanceTestCandidates,
 } from "../../acceptance/test-path";
+import { resolveAcceptanceFeatureTestPath } from "../../acceptance/test-path";
 import type { DiagnosisResult } from "../../acceptance/types";
 import { getAgent } from "../../agents/registry";
 import type { AgentAdapter } from "../../agents/types";
 import type { NaxConfig } from "../../config";
-import { resolveModelForAgent } from "../../config";
-import { loadConfigForWorkdir } from "../../config/loader";
 import { type LoadedHooksConfig, fireHook } from "../../hooks";
 import { getSafeLogger } from "../../logger";
 import type { StoryMetrics } from "../../metrics";
 import type { PipelineEventEmitter } from "../../pipeline/events";
-import { runPipeline } from "../../pipeline/runner";
-import { defaultPipeline } from "../../pipeline/stages";
-import type { AgentGetFn } from "../../pipeline/types";
-import type { PipelineContext, RoutingResult } from "../../pipeline/types";
+import type { AgentGetFn, PipelineContext } from "../../pipeline/types";
 import type { PluginRegistry } from "../../plugins";
-import { loadPRD, savePRD } from "../../prd";
-import type { PRD, UserStory } from "../../prd/types";
-import { resolveRouting } from "../../routing";
+import type { PRD } from "../../prd/types";
 import { hookCtx } from "../helpers";
 import type { StatusWriter } from "../status-writer";
+import {
+  buildResult,
+  isStubTestFile,
+  isTestLevelFailure,
+  regenerateAcceptanceTest as regenerateAcceptanceTestFn,
+} from "./acceptance-helpers";
+
+export {
+  buildResult,
+  isStubTestFile,
+  isTestLevelFailure,
+  loadAcceptanceTestContent,
+  loadSpecContent,
+  regenerateAcceptanceTest,
+  _regenerateDeps,
+} from "./acceptance-helpers";
 
 export interface AcceptanceLoopContext {
   config: NaxConfig;
@@ -75,101 +82,8 @@ export interface AcceptanceLoopResult {
   retries?: number;
 }
 
-export function isStubTestFile(content: string): boolean {
-  // Detect skeleton stubs: expect(true).toBe(false) or expect(true).toBe(true) in test bodies
-  return /expect\s*\(\s*true\s*\)\s*\.\s*toBe\s*\(\s*(?:false|true)\s*\)/.test(content);
-}
-
-/**
- * Detect test-level failure (P1-D, D2).
- *
- * Returns true when the failure is likely a test bug rather than implementation gaps:
- * - All semantic verdicts passed (overrides ratio check)
- * - Test crashed with no ACs parsed ("AC-ERROR" sentinel)
- * - More than 80% of total ACs failed
- *
- * @param failedACs - ACs that failed in this run (or number of failed ACs)
- * @param totalACs - Total ACs across all non-fix stories
- * @param semanticVerdicts - Optional semantic verdicts; when all passed, returns true
- */
-export function isTestLevelFailure(
-  failedACs: string[] | number,
-  totalACs: number,
-  semanticVerdicts?: Array<{ passed: boolean }>,
-): boolean {
-  // When all semantic verdicts passed, this is a test-level failure
-  if (semanticVerdicts && semanticVerdicts.length > 0 && semanticVerdicts.every((v) => v.passed)) {
-    return true;
-  }
-
-  const failedCount = typeof failedACs === "number" ? failedACs : failedACs.length;
-  const hasACError = Array.isArray(failedACs) && failedACs.includes("AC-ERROR");
-
-  if (hasACError) return true;
-  if (totalACs === 0) return false;
-  return failedCount / totalACs > 0.8;
-}
-
-/** Load spec.md content for AC text */
-async function loadSpecContent(featureDir?: string): Promise<string> {
-  if (!featureDir) return "";
-  const specPath = path.join(featureDir, "spec.md");
-  const specFile = Bun.file(specPath);
-  return (await specFile.exists()) ? await specFile.text() : "";
-}
-
-/**
- * Load acceptance test file content.
- *
- * When `testPaths` is provided, returns content for each per-package test file.
- * When `testPaths` is omitted, falls back to reading the configured single test file
- * from `featureDir`.
- *
- * @param featureDir - Feature directory (legacy fallback)
- * @param testPaths - Per-package test paths array (takes priority over featureDir)
- * @param configuredTestPath - Configured acceptance test path relative to featureDir
- * @returns Array of { content, path } pairs
- */
-export async function loadAcceptanceTestContent(
-  featureDir?: string,
-  testPaths?: Array<{ testPath: string; packageDir: string }>,
-  configuredTestPath?: string,
-): Promise<Array<{ content: string; path: string }>> {
-  if (!featureDir) return [];
-
-  if (testPaths && testPaths.length > 0) {
-    const results: Array<{ content: string; path: string }> = [];
-    for (const { testPath } of testPaths) {
-      const testFile = Bun.file(testPath);
-      if (await testFile.exists()) {
-        const content = await testFile.text();
-        results.push({ content, path: testPath });
-      }
-    }
-    return results;
-  }
-
-  if (!configuredTestPath) return [];
-
-  const resolvedPath = path.join(featureDir, configuredTestPath);
-  const testFile = Bun.file(resolvedPath);
-  const content = (await testFile.exists()) ? await testFile.text() : "";
-  return [{ content, path: resolvedPath }];
-}
-
-/** Build result object for loop exit */
-function buildResult(
-  success: boolean,
-  prd: PRD,
-  totalCost: number,
-  iterations: number,
-  storiesCompleted: number,
-  prdDirty: boolean,
-  failedACs?: string[],
-  retries?: number,
-): AcceptanceLoopResult {
-  return { success, prd, totalCost, iterations, storiesCompleted, prdDirty, failedACs, retries };
-}
+// isStubTestFile, isTestLevelFailure, loadSpecContent, loadAcceptanceTestContent,
+// buildResult — extracted to acceptance-helpers.ts (re-exported above)
 
 export const _acceptanceLoopDeps = {
   getAgent,
@@ -186,7 +100,7 @@ export const _acceptanceLoopDeps = {
       language: ctx.config.project?.language,
     });
     if (!testPath) return "no_test_file";
-    const regenerated = await regenerateAcceptanceTest(testPath, acceptanceContext, previousFailure);
+    const regenerated = await regenerateAcceptanceTestFn(testPath, acceptanceContext, previousFailure);
     if (!regenerated) return "failed";
     const { acceptanceStage } = await import("../../pipeline/stages/acceptance");
     const result = await acceptanceStage.execute(acceptanceContext);
@@ -194,212 +108,8 @@ export const _acceptanceLoopDeps = {
   },
 };
 
-/** Injectable dependencies for regenerateAcceptanceTest — allows tests to mock I/O without real disk or git. */
-export const _regenerateDeps = {
-  spawnGitDiff: async (workdir: string, gitRef: string): Promise<string> => {
-    const proc = Bun.spawn(["git", "diff", "--name-only", gitRef], {
-      cwd: workdir,
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-    const [, stdout] = await Promise.all([proc.exited, new Response(proc.stdout).text()]);
-    return stdout.trim();
-  },
-  readFile: async (filePath: string): Promise<string> => Bun.file(filePath).text(),
-  acceptanceSetupExecute: async (ctx: PipelineContext): Promise<void> => {
-    const { acceptanceSetupStage } = await import("../../pipeline/stages/acceptance-setup");
-    await acceptanceSetupStage.execute(ctx);
-  },
-};
-
-/** Generate and add fix stories to PRD */
-async function generateAndAddFixStories(
-  ctx: AcceptanceLoopContext,
-  failures: { failedACs: string[]; testOutput: string },
-  prd: PRD,
-): Promise<FixStory[] | null> {
-  const logger = getSafeLogger();
-  const agent = (ctx.agentGetFn ?? _acceptanceLoopDeps.getAgent)(ctx.config.autoMode.defaultAgent);
-  if (!agent) {
-    logger?.error("acceptance", "Agent not found, cannot generate fix stories");
-    return null;
-  }
-  const modelDef = resolveModelForAgent(
-    ctx.config.models,
-    ctx.config.autoMode.defaultAgent,
-    ctx.config.analyze.model,
-    ctx.config.autoMode.defaultAgent,
-  );
-  const testFilePath = ctx.featureDir
-    ? resolveAcceptanceFeatureTestPath(ctx.featureDir, ctx.config.acceptance.testPath, ctx.config.project?.language)
-    : undefined;
-  const fixStories = await generateFixStories(agent, {
-    failedACs: failures.failedACs,
-    testOutput: failures.testOutput,
-    prd,
-    specContent: await loadSpecContent(ctx.featureDir),
-    workdir: ctx.workdir,
-    modelDef,
-    config: ctx.config,
-    testFilePath,
-    timeoutMs: ctx.config.acceptance?.timeoutMs,
-  });
-  if (fixStories.length === 0) {
-    logger?.error("acceptance", "Failed to generate fix stories");
-    return null;
-  }
-  logger?.info("acceptance", `Generated ${fixStories.length} fix stories`);
-  for (const fixStory of fixStories) {
-    const userStory = convertFixStoryToUserStory(fixStory);
-    prd.userStories.push(userStory);
-    logger?.debug("acceptance", `Fix story added: ${userStory.id}: ${userStory.title}`);
-  }
-  return fixStories;
-}
-
-/** Execute a single fix story through the pipeline */
-async function executeFixStory(
-  ctx: AcceptanceLoopContext,
-  story: UserStory,
-  prd: PRD,
-  iterations: number,
-): Promise<{ success: boolean; cost: number; metrics?: StoryMetrics[] }> {
-  const logger = getSafeLogger();
-  const routing = await resolveRouting(story, ctx.config, ctx.pluginRegistry);
-  logger?.info("acceptance", `Starting fix story: ${story.id}`, { storyId: story.id, storyTitle: story.title });
-  await fireHook(
-    ctx.hooks,
-    "on-story-start",
-    hookCtx(ctx.feature, {
-      storyId: story.id,
-      model: routing.modelTier,
-      agent: ctx.config.autoMode.defaultAgent,
-      iteration: iterations,
-    }),
-    ctx.workdir,
-  );
-  // PKG: resolve per-package effective config for fix stories (same as iteration-runner)
-  const fixEffectiveConfig = story.workdir
-    ? await loadConfigForWorkdir(join(ctx.workdir, ".nax", "config.json"), story.workdir)
-    : ctx.config;
-  const fixContext: PipelineContext = {
-    config: fixEffectiveConfig,
-    rootConfig: ctx.config,
-    prd,
-    story,
-    stories: [story],
-    routing: routing as RoutingResult,
-    projectDir: ctx.workdir,
-    workdir: story.workdir ? join(ctx.workdir, story.workdir) : ctx.workdir,
-    featureDir: ctx.featureDir,
-    hooks: ctx.hooks,
-    plugins: ctx.pluginRegistry,
-    storyStartTime: new Date().toISOString(),
-    agentGetFn: ctx.agentGetFn,
-  };
-  const result = await runPipeline(defaultPipeline, fixContext, ctx.eventEmitter);
-  logger?.info("acceptance", `Fix story ${story.id} ${result.success ? "passed" : "failed"}`);
-  return {
-    success: result.success,
-    cost: result.context.agentResult?.estimatedCost || 0,
-    metrics: result.context.storyMetrics,
-  };
-}
-
-/**
- * Back up and regenerate the acceptance test file (P1-D, D2).
- *
- * Steps:
- * 1. Copy configured acceptance test file → <file>.bak
- * 2. Delete configured acceptance test file
- * 3. Re-run acceptance-setup to generate fresh test
- *
- * @returns true if regeneration succeeded, false otherwise
- */
-export async function regenerateAcceptanceTest(
-  testPath: string,
-  acceptanceContext: PipelineContext,
-  previousFailure?: string,
-): Promise<boolean> {
-  const logger = getSafeLogger();
-  const bakPath = `${testPath}.bak`;
-
-  const content = await Bun.file(testPath).text();
-  await Bun.write(bakPath, content);
-  logger?.info("acceptance", `Backed up acceptance test -> ${bakPath}`);
-
-  const { unlink } = await import("node:fs/promises");
-  await unlink(testPath);
-
-  // Delete acceptance-meta.json so acceptanceSetupExecute cannot reuse the
-  // fingerprint of the (now-deleted) test and is forced to regenerate.
-  if (acceptanceContext.featureDir) {
-    const metaPath = path.join(acceptanceContext.featureDir, "acceptance-meta.json");
-    try {
-      await unlink(metaPath);
-    } catch {
-      // missing meta is fine — setup will treat it as not-yet-generated
-    }
-  }
-
-  // Collect implementation context from git diff when storyGitRef is available
-  let implementationContext: Array<{ path: string; content: string }> | undefined;
-  const storyGitRef = acceptanceContext.storyGitRef;
-  const workdir = acceptanceContext.workdir;
-
-  if (storyGitRef) {
-    try {
-      const diffOutput = await _regenerateDeps.spawnGitDiff(workdir, storyGitRef);
-      const changedFiles = diffOutput
-        .split("\n")
-        .map((f) => f.trim())
-        .filter((f) => f.length > 0);
-
-      const MAX_BYTES = 50 * 1024;
-      let totalBytes = 0;
-      const entries: Array<{ path: string; content: string }> = [];
-
-      for (const file of changedFiles) {
-        if (totalBytes >= MAX_BYTES) break;
-        const filePath = path.join(workdir, file);
-        try {
-          const fileContent = await _regenerateDeps.readFile(filePath);
-          const remaining = MAX_BYTES - totalBytes;
-          const trimmed = fileContent.length > remaining ? fileContent.slice(0, remaining) : fileContent;
-          entries.push({ path: file, content: trimmed });
-          totalBytes += trimmed.length;
-        } catch {
-          // skip unreadable files
-        }
-      }
-
-      if (entries.length > 0) {
-        implementationContext = entries;
-      }
-    } catch {
-      // git diff failed — proceed without implementation context
-    }
-  }
-
-  const contextForSetup: PipelineContext & {
-    implementationContext?: Array<{ path: string; content: string }>;
-    previousFailure?: string;
-  } = {
-    ...acceptanceContext,
-    ...(implementationContext ? { implementationContext } : {}),
-    ...(previousFailure ? { previousFailure } : {}),
-  };
-
-  await _regenerateDeps.acceptanceSetupExecute(contextForSetup as PipelineContext);
-
-  if (!(await Bun.file(testPath).exists())) {
-    logger?.error("acceptance", "Acceptance test regeneration failed — manual intervention required");
-    return false;
-  }
-
-  logger?.info("acceptance", "Acceptance test regenerated successfully");
-  return true;
-}
+// _regenerateDeps, regenerateAcceptanceTest, generateAndAddFixStories, executeFixStory
+// — extracted to acceptance-helpers.ts or deleted (dead code)
 
 export interface FixRoutingOptions {
   ctx: AcceptanceLoopContext;
@@ -663,7 +373,11 @@ export async function runFixRouting(options: FixRoutingOptions): Promise<FixRout
     }
 
     const failureContext = `Diagnosis: ${diagnosis.reasoning}\n\nFailing test output:\n${failures.testOutput}`;
-    const regenerated = await regenerateAcceptanceTest(testPath, acceptanceContext as PipelineContext, failureContext);
+    const regenerated = await regenerateAcceptanceTestFn(
+      testPath,
+      acceptanceContext as PipelineContext,
+      failureContext,
+    );
 
     logger?.info("acceptance.test-regen", "Test regeneration completed", {
       outcome: regenerated ? "success" : "failure",
@@ -777,7 +491,7 @@ export async function runFixRouting(options: FixRoutingOptions): Promise<FixRout
     }
 
     const bothFailureContext = `Diagnosis: ${diagnosis.reasoning}\n\nFailing test output:\n${failures.testOutput}`;
-    const regenerated = await regenerateAcceptanceTest(
+    const regenerated = await regenerateAcceptanceTestFn(
       testPath,
       acceptanceContext as PipelineContext,
       bothFailureContext,
@@ -803,11 +517,11 @@ export async function runAcceptanceLoop(ctx: AcceptanceLoopContext): Promise<Acc
   const maxRetries = ctx.config.acceptance.maxRetries;
 
   let acceptanceRetries = 0;
-  let prd = ctx.prd;
+  const prd = ctx.prd;
   let totalCost = ctx.totalCost;
-  let iterations = ctx.iterations;
-  let storiesCompleted = ctx.storiesCompleted;
-  let prdDirty = false;
+  const iterations = ctx.iterations;
+  const storiesCompleted = ctx.storiesCompleted;
+  const prdDirty = false;
 
   logger?.info("acceptance", "All stories complete, running acceptance validation");
 
@@ -951,7 +665,7 @@ export async function runAcceptanceLoop(ctx: AcceptanceLoopContext): Promise<Acc
       });
       if (testPath) {
         const testLevelFailureContext = `Test-level failure: ${failures.failedACs.length}/${totalACs} ACs failed.\n\nFailing test output:\n${failures.testOutput}`;
-        const regenerated = await regenerateAcceptanceTest(testPath, acceptanceContext, testLevelFailureContext);
+        const regenerated = await regenerateAcceptanceTestFn(testPath, acceptanceContext, testLevelFailureContext);
         if (!regenerated) {
           return buildResult(
             false,
@@ -1002,49 +716,8 @@ export async function runAcceptanceLoop(ctx: AcceptanceLoopContext): Promise<Acc
       );
     }
 
-    // Fallback: generate and add fix stories (legacy path)
-    logger?.info("acceptance", "Generating fix stories...");
-    const fixStories = await generateAndAddFixStories(ctx, failures, prd);
-    if (!fixStories) {
-      return buildResult(
-        false,
-        prd,
-        totalCost,
-        iterations,
-        storiesCompleted,
-        prdDirty,
-        failures.failedACs,
-        acceptanceRetries,
-      );
-    }
-
-    await savePRD(prd, ctx.prdPath);
-    prdDirty = true;
-
-    // Execute fix stories
-    logger?.info("acceptance", "Running fix stories...");
-    for (const fixStory of fixStories) {
-      const userStory = prd.userStories.find((s) => s.id === fixStory.id);
-      if (!userStory || userStory.status !== "pending") continue;
-
-      iterations++;
-      const result = await executeFixStory(ctx, userStory, prd, iterations);
-      prd = await loadPRD(ctx.prdPath); // Reload to get updated PRD
-
-      if (result.success) {
-        storiesCompleted++;
-        totalCost += result.cost;
-        // Mutates ctx.allStoryMetrics in place — ctx.allStoryMetrics is the same array reference
-        // as options.allStoryMetrics in runner-completion.ts, so fix-story metrics are visible
-        // to shouldSkipDeferredRegression when handleRunCompletion is called after this loop.
-        if (result.metrics) ctx.allStoryMetrics.push(...result.metrics);
-      }
-
-      await savePRD(prd, ctx.prdPath);
-      prdDirty = true;
-    }
-
-    logger?.info("acceptance", "Re-running acceptance tests...");
+    // Legacy fallback path removed — strategy is always "diagnose-first" | "implement-only"
+    // (enforced by Zod enum in AcceptanceFixConfigSchema). The if-block above is always entered.
   }
 
   return buildResult(false, prd, totalCost, iterations, storiesCompleted, prdDirty);
